adding The notebook file and add ACP to main.py
