--- PCA_DEV2_branch.py
@@ -1,85 +1,0 @@
-from sklearn.datasets import fetch_20newsgroups
-from sklearn.metrics.cluster import normalized_mutual_info_score, adjusted_rand_score
-from sentence_transformers import SentenceTransformer
-from sklearn.decomposition import PCA
-from sklearn.cluster import KMeans
-import numpy as np
-
-'''
-Variables:
----------
-
-corpus : list of documents
-embeddings : documents embeddings of size NxM (N : number of documents, M : embedding dimension)
-red_emd : reduced embeddings matrix using dimentionality reduction
-k : number of clusters
-labels : documents labels
-pred : list of clustering predicted clusters
-
-'''
-
-def dim_red(mat, p):
-    '''
-    Perform dimensionality reduction
-
-    Input:
-    -----
-        mat : NxM list
-        p : number of dimensions to keep
-    Output:
-    ------
-        red_mat : NxP list such that p<<m
-    '''
-
-    # Apply PCA for dimensionality reduction
-    pca = PCA(n_components=p)
-    red_mat = pca.fit_transform(mat)
-
-    red_mat = mat[:,:p]
-
-    return red_mat
-
-
-
-def clust(mat, k):
-    '''
-    Perform clustering
-
-    Input:
-    -----
-        mat : input list
-        k : number of cluster
-    Output:
-    ------
-        pred : list of predicted labels
-    '''
-    kmeans = KMeans(n_clusters=k)
-    kmeans.fit(mat)
-    pred = kmeans.labels_
-
-    return pred
-    #pred = np.random.randint(k, size=len(corpus))
-
-    #return pred
-
-# import data
-ng20 = fetch_20newsgroups(subset='test')
-corpus = ng20.data[:2000]
-labels = ng20.target[:2000]
-k = len(set(labels))
-
-# embedding
-model = SentenceTransformer('paraphrase-MiniLM-L6-v2')
-embeddings = model.encode(corpus)
-
-# perform dimentionality reduction
-red_emb = dim_red(embeddings, 20)
-# perform clustering
-pred = clust(red_emb, k)
-
-# evaluate clustering results
-nmi_score = normalized_mutual_info_score(pred,labels)
-ari_score = adjusted_rand_score(pred,labels)
-
-print(f'NMI: {nmi_score:.2f} \nARI: {ari_score:.2f}')
-
